Ryan casts an int and catches a fish, someone buy him a pint! merge to 5092
git-svn-id: http://mu.svn.wordpress.org/trunk@933 7be80a69-a1ef-0310-a953-fb0f7c49ff36

+++ wp-admin/import/wordpress.php
@@ -157,7 +157,7 @@ class WP_Import {
 			return;
 		}
 		$this->file = $file['file'];
-		$this->id = $file['id'];
+		$this->id = (int) $file['id'];
 
 		$this->get_entries();
 		$this->wp_authors_form();
@@ -184,7 +184,7 @@ class WP_Import {
 			if ( empty($parent) )
 				$category_parent = '0';
 			else
-				$category_parent = (int) category_exists($parent);
+				$category_parent = category_exists($parent);
 
 			$catarr = compact('category_nicename', 'category_parent', 'posts_private', 'links_private', 'posts_private', 'cat_name');
